changes + bump react-icons
