+++ frontend/public/env_template.js
@@ -4,4 +4,7 @@ window.$env = {
   RECAPTCHA_SITEKEY: "",
   // b) public deployment
   // RECAPTCHA_SITEKEY: "",
+
+  // set variable to "local" for internal development, "prod" for public deployment
+  STAGE: "prod",
 };
