+++ frontend/src/layouts/AppMain.jsx
@@ -3,7 +3,12 @@ import { useRoutes, Outlet, useLocation } from "react-router-dom";
 import PropTypes from "prop-types";
 import { ErrorBoundary } from "react-error-boundary";
 import { Row, Col } from "reactstrap";
-import { FallBackLoading, ErrorAlert } from "@certego/certego-ui";
+import {
+  FallBackLoading,
+  ErrorAlert,
+  useToastr,
+  Toaster,
+} from "@certego/certego-ui";
 
 // wrapper
 import withAuth from "../wrappers/withAuth";
@@ -40,7 +45,12 @@ ErrorHandler.propTypes = {
   error: PropTypes.object.isRequired,
 };
 
+// constants
+const selector = (state) => state.toasts;
+
 function Layout() {
+  // consume store
+  const toasts = useToastr(selector);
   return (
     <>
       <AppHeader />
@@ -49,6 +59,12 @@ function Layout() {
           <Outlet />
         </ErrorBoundary>
       </main>
+      {/* Toasts */}
+      <section className="fixed-bottom" id="app-toasts">
+        {toasts.map((tProps) => (
+          <Toaster key={tProps.id} {...tProps} />
+        ))}
+      </section>
     </>
   );
 }
