+++ frontend/src/styles/App.scss
@@ -105,29 +105,13 @@ section.fixed-bottom {
   overflow-y: auto;
 }
 
-.twitter-follow-button {
-  /* stylelint-disable scss/at-extend-no-missing-placeholder */
-  @extend .btn;
-  @extend .btn-xs;
-  /* stylelint-enable scss/at-extend-no-missing-placeholder */
-
-  background-color: #1da1f2;
-  color: $light;
-  font-weight: 450;
-
-  &:hover {
-    background-color: shade-color(#1da1f2, 20%);
-    color: $light !important;
-  }
-}
-
 .btn-social {
   /* stylelint-disable scss/at-extend-no-missing-placeholder */
   @extend .btn;
   @extend .btn-xs;
   /* stylelint-enable scss/at-extend-no-missing-placeholder */
 
-  background-color: #1da1f2;
+  background-color: $dark;
   color: $light;
   font-weight: 450;
